MonadBaseControl and other instances for SafeT
A main difference between SafeT's implementation of MonadBaseControl,
and that of ResourceT: ResourceT allows a reference counting mechanism
so that multiple threads can share the same resource environment, but
where resources are only freed once the last thread has finished
executing.

I'm not sure yet if we need this behavior, so I haven't coded in
support (though it will be fairly easy to add later). Instead, and as
before, runSafeT governs the lifetime of resources. If you spawn threads
in a runSafeT block and they allocate resources, those resources are
freed when runSafeT ends, even if those threads are still running (which
means they could now reference freed resources). If the child threads
attempt to acquire or release resources beyond that point, it raises an
error.

Therefore, the only safe way to mix forkIO with SafeT (via
MonadBaseControl) is to ensure that the thread executing runSafeT either
waits until all other threads have been completed before allowing
runSafeT to complete, or cancels all those threads before it exits.